feat(monitor): relocate QNN helpers → session/monitor/qnn/
Moves csv_parser.py, qhas_parser.py, viewer.py + test fixtures from
optracing/qnn/ to session/monitor/qnn/, the canonical home for
QNNMonitor support code. Old paths retained as backward-compat shims.

Also updates test_integration.py fixture path and test_qnn_profiler.py
import + patch targets to reference the new module location.

Constraint: Old import paths must remain importable for any code already
referencing optracing.qnn.*
Scope-risk: narrow
Confidence: high
