+++ src/winml/modelkit/session/monitor/qnn/__init__.py
@@ -1,0 +1,5 @@
+# -------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License.
+# --------------------------------------------------------------------------
+"""QNN-specific helpers for QNNMonitor: CSV parser, QHAS parser, viewer shell-out."""
