+++ tests/unit/session/monitor/qnn/fixtures/qhas_resnet50.json
